Added Open, Save As and New file menus

+++ Deya/src/Deya/Scene/SceneSerializer.cpp
@@ -124,12 +124,12 @@ namespace Deya
             out << YAML::Key << "Camera" << YAML::Value;
             out << YAML::BeginMap; // Camera
             out << YAML::Key << "ProjectionType" << YAML::Value << (int) camera.GetProjectionType();
-            out << YAML::Key << "PerspectiveFOV" << YAML::Value << (int) camera.GetPerspectiveVerticalFOV();
-            out << YAML::Key << "PerspectiveNear" << YAML::Value << (int) camera.GetPerspectiveNearClip();
-            out << YAML::Key << "PerspectiveFar" << YAML::Value << (int) camera.GetPerspectiveFarClip();
-            out << YAML::Key << "OrthographicSize" << YAML::Value << (int) camera.GetOrthographicSize();
-            out << YAML::Key << "OrthographicNear" << YAML::Value << (int) camera.GetOrthographicNearClip();
-            out << YAML::Key << "OrthographicFar" << YAML::Value << (int) camera.GetOrthographicFarClip();
+            out << YAML::Key << "PerspectiveFOV" << YAML::Value << camera.GetPerspectiveVerticalFOV();
+            out << YAML::Key << "PerspectiveNear" << YAML::Value << camera.GetPerspectiveNearClip();
+            out << YAML::Key << "PerspectiveFar" << YAML::Value << camera.GetPerspectiveFarClip();
+            out << YAML::Key << "OrthographicSize" << YAML::Value << camera.GetOrthographicSize();
+            out << YAML::Key << "OrthographicNear" << YAML::Value << camera.GetOrthographicNearClip();
+            out << YAML::Key << "OrthographicFar" << YAML::Value << camera.GetOrthographicFarClip();
             out << YAML::EndMap; // Camera
 
             out << YAML::Key << "Primary" << YAML::Value << cameraComponent.Primary;
